fix(sse): address Centaur review — round 2
- Fix heartbeat Map mutation during iteration (collect-then-remove)
- Fix TDZ risk: declare overviewEmitter before broadcast closures
- Remove non-null assertion on session.sessionId
- Skip empty listener sets in EventBus createSource()
- Move SessionActivity types to @mitzo/protocol (single source of truth)
